implement beginning of program status thing
