+++ src/main.rs
@@ -1,5 +1,6 @@
 mod audio;
-pub mod json_processor;
+mod keybind;
+mod json_processor;
 
 use anyhow::{Context, Error};
 use clap::Parser;
@@ -13,7 +14,6 @@ use std::io::Write;
 use tokio::{
     sync::mpsc::{Sender, channel},
     net::TcpStream,
-    time::{sleep, Duration},
     task,
 };
 use tokio_tungstenite::{
@@ -22,7 +22,6 @@ use tokio_tungstenite::{
     WebSocketStream,
     tungstenite::protocol::Message,
 };
-use device_query::{DeviceQuery, DeviceState, Keycode};
 
 /// A miniaturized version of FE2.IO written in Rust, and independent of any web browsers.
 #[derive(Parser, Clone)]
@@ -72,7 +71,7 @@ async fn main() -> Result<(), Error> {
     // Spawn separate task for handling websocket events
     task::spawn(websocket_loop(tx.clone(), read));
 
-    keybind_listen(tx, volume_tx, volume).await?;
+    keybind::keybind_listen(tx, volume_tx, volume).await?;
     Ok(())
 }
 
@@ -98,44 +97,5 @@ async fn websocket_loop(tx: Sender<String>, mut read: SplitStream<WebSocketStrea
         let data = message?.into_text()?;
         println!("Got message {}", data);
         json_processor::process_data(&data, &tx).await?;
-    }
-}
-
-async fn keybind_listen(tx: Sender<String>, volume_tx: Sender<f32>, mut volume: f32) -> Result<(), Error> {
-    let device_state = DeviceState::new();
-    let mut initial_volume = volume;
-    let mut muted = false;
-    loop {
-        let keys: Vec<Keycode> = device_state.get_keys();
-        match keys {
-            keys if keys.contains(&Keycode::Equal) => {
-                volume = (((volume * 100.0) + 5.0).min(100.0).round()) / 100.0;
-                volume_tx.send(volume).await?; // Increase volume
-                tx.send("volume".to_owned()).await?; // Send volume event to audio loop
-                muted = false;
-            },
-            keys if keys.contains(&Keycode::Minus) => {
-                volume = (((volume * 100.0) - 5.0).max(0.0).round()) / 100.0;
-                volume_tx.send(volume).await?; // Lower volume
-                tx.send("volume".to_owned()).await?;
-                muted = false;
-            },
-            keys if keys.contains(&Keycode::Grave) => { // this key `
-                if muted {
-                    volume = initial_volume;
-                    volume_tx.send(volume).await?; // Unmute (sets volume to value before muted)
-                    tx.send("volume".to_owned()).await?;
-                    muted = false;
-                } else {
-                    initial_volume = volume; // Save initial volume in variable
-                    volume = 0.0;
-                    volume_tx.send(volume).await?; // Mute
-                    tx.send("volume".to_owned()).await?;
-                    muted = true;
-                }
-            },
-            _ => (),
-        }
-        sleep(Duration::from_millis(100)).await; // sleep for 100 ms to avoid maxing out the cpu
     }
 }
